Create Linked list.py
Basic operations in linked list:

Insertion- Adds an element at the beginning of the list
Deletion - Deletes an element at the beginning of the list
Display- Display the complete list.
Seach- searches and element using the given key


Linked list - insertion Operation
Adding a new node in linked list is a more than one step activity.We shall learn this with diagrams here. First ,create a node using the the same structure and find the location where it has to beinserted